+++ MoMoney.Core/Converters/EmptyToVisibleConverter.cs
@@ -5,7 +5,7 @@ namespace MoMoney.Core.Converters;
 /// <summary>
 /// Changes visibility of button based on whether the corresponding value is empty.
 /// </summary>
-class EmptyToVisibleConverter : IValueConverter
+public class EmptyToVisibleConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {