+++ MoMoney.Core/Converters/NoneToVisibleConverter.cs
@@ -6,7 +6,7 @@ namespace MoMoney.Core.Converters;
 /// <summary>
 /// Changes visibility of button based on whether the corresponding value is None.
 /// </summary>
-class NoneToVisibleConverter : IValueConverter
+public class NoneToVisibleConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {