+++ MoMoney.Core/Converters/ShowValuePercentConverter.cs
@@ -6,7 +6,7 @@ namespace MoMoney.Core.Converters;
 /// <summary>
 /// Converts value to "*.**%" if Utilities.ShowValue is true
 /// </summary>
-class ShowValuePercentConverter : IValueConverter
+public class ShowValuePercentConverter : IValueConverter
 {
     public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
     {