+++ MoMoney.Tests/UnitTest1.cs
@@ -1,12 +1,11 @@
 
-namespace MoMoney.Tests
+namespace MoMoney.Tests;
+
+public class UnitTest1
 {
-    public class UnitTest1
+    [Fact]
+    public void Test1()
     {
-        [Fact]
-        public void Test1()
-        {
-            
-        }
+        
     }
 }